dialysis_scheduler/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
flask-wtf==1.2.1
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3.post1